Frontends/VBoxManage: 'VBoxManage internalcommands createrawvmdk' no longer works after the implementation was moved to the API as part of xTracker #9224.
Follow-up build fix to address a left-behind orphaned function
(appendPartDesc()) whose only caller in VBoxInternalManage.cpp was
removed (flagged as 'defined but not used' by [-Werror=unused-function]).
Also, avoid assigning a size_t to an int which is flagged as 'possible
loss of data' by the Windows compiler.

bugref:9224 ticketref:21125